Add files via upload
In this Python project, I have compiled a dataset of world cup matches from 1930 to 2018, demonstrating the countries that have played the most and won the most football matches and world cups. During the same interval, # This data will be used to determine the country that has participated in the most World Cup matches, the top 10 countries that have participated in the most World Cup matches, to create a bar chart for the top five countries, and to determine the countries with the highest number of goals and winning teams from 1930 to 2018. Throughout the duration of the project, I have learned numerous functions, and I continue to learn and enjoy the Python programming language.