Add files via upload
The End-to-End Bluebook Bulldozer Price Regression project is a machine learning project that aims to predict the sale price of bulldozers based on their characteristics, such as model type, age, usage, and configuration. The project uses a dataset of historical bulldozer sales, which includes information on their features and sale prices.

The project involves several steps, including data cleaning and preprocessing, feature engineering, model selection and training, and model evaluation.