Add files via upload
The K-Nearest Neighbors (K-NN) algorithm operates by identifying the closest data points to a target value. To begin, a value for 'k' is specified, denoting the number of neighboring points to consider. For instance, with k=5, the algorithm locates the five closest data points to the target and computes their distances using two methods: Euclidean and Manhattan.

It's important to note that the value of 'k' should be an odd number. Before implementing the K-NN algorithm, data scaling is recommended. Common methods for scaling include Z-score normalization or standard scaling.

Below is an example of code employing the K-NN algorithm to classify malignant and benign cancers. Rather than randomly selecting 'k', this code assists in determining the optimal 'k' value to minimize error rates.